refactor: move inputs.ts and outputs.ts to _shared module
