+++ packages/runtime/src/_shared/index.ts
@@ -1,3 +1,5 @@
 // Copyright (c) 2020-2022 Climate Interactive / New Venture Fund
 
 export type { InputVarId, OutputVarId, OutputVarSpec } from './types'
+export * from './inputs'
+export * from './outputs'
