+++ packages/runtime/src/index.ts
@@ -1,6 +1,6 @@
 // Copyright (c) 2020-2022 Climate Interactive / New Venture Fund
 
-export type { InputVarId, OutputVarId, OutputVarSpec } from './_shared'
+export * from './_shared'
 export * from './wasm-model'
 export * from './model-runner'
 export * from './model-scheduler'
