+++ packages/runtime/src/model-runner/index.ts
@@ -1,7 +1,5 @@
 // Copyright (c) 2020-2022 Climate Interactive / New Venture Fund
 
-export * from './inputs'
-export * from './outputs'
 export * from './model-runner'
 
 /** @hidden This is not part of the public API; exposed only for use in testing tools. */
